fix: rename zustand typos.
