+++ src/App.tsx
@@ -2,18 +2,18 @@ import logo from './logo.svg'
 import './App.css'
 import {EstadoGlobal} from './components/EstadoGlobal';
 import {TesteContextApiProvider} from './store/ContexApi';
-import ProviderContexZustend from './store/ContextZuztandStore';
+import ProviderContexZustand from './store/ContextZustandStore';
 
 function App() {
   return (
     <TesteContextApiProvider>
-      <ProviderContexZustend initialNumber={2}>
+      <ProviderContexZustand initialNumber={2}>
         <div className="App">
           <header className="App-header">
              <EstadoGlobal />
           </header>
         </div>
-      </ProviderContexZustend>
+      </ProviderContexZustand>
     </TesteContextApiProvider>
 
   )
