+++ src/components/EstadoGlobal/useEstadoGlobalController.ts
@@ -1,7 +1,7 @@
 import {useState} from 'react';
 import {useTesteContextApi} from '../../store/ContexApi';
 import {useZustandStore} from '../../store/ZustandStore';
-import {useStoreContextZustand} from '../../store/ContextZuztandStore';
+import {useStoreContextZustand} from '../../store/ContextZustandStore';
 
 interface EstadoGlobalControllerProps {
   contagem: number;
